Try to sort common entries first.
It already did it when searching by character (i.e. would try to show you
common readings before less common readings for same character), but not it
also does it when searching by pinyin. For example, if you search for "ren"
the first result will be "人" instead of something obscure.